Moved test utils to test's init; added isort config
