+++ hdlcc/tests/test_vhdl_parser.py
@@ -29,13 +29,14 @@ import six
 
 from nose2.tools import such  # type: ignore
 
+from hdlcc.tests import assertCountEqual, assertSameFile, writeListToFile
+
 from hdlcc.parsers.elements.dependency_spec import DependencySpec
 from hdlcc.parsers.elements.design_unit import DesignUnitType, VhdlDesignUnit
 from hdlcc.parsers.elements.identifier import Identifier
 from hdlcc.parsers.vhdl_parser import VhdlParser
 from hdlcc.path import Path
 from hdlcc.serialization import StateEncoder, jsonObjectHook
-from hdlcc.tests.utils import assertCountEqual, assertSameFile, writeListToFile
 
 _logger = logging.getLogger(__name__)
 
@@ -111,7 +112,7 @@ with such.A("VHDL source file object") as it:
             if os.path.exists(_FILENAME):
                 os.remove(_FILENAME)
 
-        @it.should("parse a file without errors")  # type: ignore
+        @it.should("parse a file without errors")
         def test():
             it.source = VhdlParser(Path(_FILENAME))
