+++ setup.cfg
@@ -11,3 +11,11 @@ python_version = 3.7
 follow_imports = silent
 warn_unused_configs = true
 warn_unused_ignores = true
+
+[isort]
+forced_separate=nose2,hdlcc.tests,hdlcc
+multi_line_output=3
+include_trailing_comma=True
+force_grid_wrap=0
+use_parentheses=True
+line_length=88
